change the scan modal to standalone page
