--- csc-courier-ssapp/code/pages/scan_shipment.html
@@ -1,12 +1,0 @@
-<webc-container controller="ScanShipmentController">
-  <main class="full-flex-column">
-    <section>
-      <webc-template template="header-template"></webc-template>
-      <webc-template template="dashboard-menu"></webc-template>
-    </section>
-
-
-    scan Shipment
-
-  </main>
-</webc-container>
